% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survAudit.R
\name{.plot_calibration}
\alias{.plot_calibration}
\title{Plot calibration diagnostics}
\usage{
.plot_calibration(cal, col_point, col_ref, alpha_pt)
}
\arguments{
\item{cal}{Calibration diagnostics list.}

\item{col_point}{Point colour.}

\item{col_ref}{Reference line colour.}

\item{alpha_pt}{Point alpha.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot calibration diagnostics
}
\keyword{internal}
