% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survAudit.R
\name{.plot_functional}
\alias{.plot_functional}
\title{Plot functional form diagnostics}
\usage{
.plot_functional(ff, col_point, col_smooth, col_ref, alpha_pt, vars = NULL)
}
\arguments{
\item{ff}{Functional form diagnostics list.}

\item{col_point}{Point colour.}

\item{col_smooth}{Smooth line colour.}

\item{col_ref}{Reference line colour.}

\item{alpha_pt}{Point alpha.}

\item{vars}{Optional character or integer vector of covariates to plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot functional form diagnostics
}
\keyword{internal}
