% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survAudit.R
\name{.plot_gof}
\alias{.plot_gof}
\title{Plot goodness-of-fit diagnostics}
\usage{
.plot_gof(gof, col_point, col_ref, alpha_pt)
}
\arguments{
\item{gof}{Goodness-of-fit diagnostics list.}

\item{col_point}{Point colour.}

\item{col_ref}{Reference line colour.}

\item{alpha_pt}{Point alpha.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot goodness-of-fit diagnostics
}
\keyword{internal}
