% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survAudit.R
\name{.plot_influence}
\alias{.plot_influence}
\title{Plot influence diagnostics}
\usage{
.plot_influence(inf, col_point, col_smooth, col_ref, col_thresh, alpha_pt)
}
\arguments{
\item{inf}{Influence diagnostics list.}

\item{col_point}{Point colour.}

\item{col_smooth}{Smooth line colour (unused here).}

\item{col_ref}{Reference line colour.}

\item{col_thresh}{Threshold line colour.}

\item{alpha_pt}{Point alpha.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot influence diagnostics
}
\keyword{internal}
