% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survAudit.R
\name{.plot_outliers}
\alias{.plot_outliers}
\title{Plot outlier diagnostics}
\usage{
.plot_outliers(ol, col_point, col_ref, alpha_pt)
}
\arguments{
\item{ol}{Outlier diagnostics list.}

\item{col_point}{Point colour.}

\item{col_ref}{Reference line colour.}

\item{alpha_pt}{Point alpha.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot outlier diagnostics
}
\keyword{internal}
