% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.survAudit.R
\name{.plot_ph}
\alias{.plot_ph}
\title{Plot PH diagnostics (Schoenfeld residuals)}
\usage{
.plot_ph(ph, col_point, col_smooth, col_ref, alpha_pt)
}
\arguments{
\item{ph}{PH diagnostics list from the survAudit object.}

\item{col_point}{Point colour.}

\item{col_smooth}{Smooth line colour.}

\item{col_ref}{Reference line colour.}

\item{alpha_pt}{Point alpha.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot PH diagnostics (Schoenfeld residuals)
}
\keyword{internal}
